area-rit/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.36.8
botocore==1.36.8
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Django==5.1.3
django-allauth==65.2.0
django-storages==1.14.4
fonttools==4.55.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
jmespath==1.0.1
josepy==1.15.0
hdf5plugin==5.1.0
kiwisolver==1.4.7
matplotlib==3.9.2
minio==7.2.12
mozilla-django-oidc==4.0.1
nexusformat==1.0.8
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
s3transfer==0.11.2
scipy==1.15.2
six==1.16.0
sqlparse==0.5.2
tifffile==2025.3.30
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.5.0
